// src/types/pages/floorDashboard.ts
export interface IDragButton {
  No: number
  Type: string
  FloorItemNo: number
  No_Type: string
  Name: string
  Position: {
    x: number
    y: number
  }
  Status: number
  Alert: number
  Icon: IconDefinition
}